+++ lib/trailblazer/autoloading.rb
@@ -1,5 +1,5 @@
 Trailblazer::Operation.class_eval do
   autoload :Controller, 'trailblazer/operation/controller'
-  autoload :Responder,  'trailblazer/operation/responder'
   autoload :CRUD,       'trailblazer/operation/crud'
+  autoload :Responder,  'trailblazer/operation/responder'
 end