+++ test/rails/fake_app/models.rb
@@ -6,6 +6,10 @@ class Band < ActiveRecord::Base
   has_many :songs
 end
 
+class Tenant < ActiveRecord::Base
+  self.table_name = 'public.tenants'
+end
+
 # migrations
 class CreateAllTables < ActiveRecord::Migration
   def self.up
@@ -19,7 +23,12 @@ class CreateAllTables < ActiveRecord::Migration
       t.string :name
       t.string :locality
     end
+
+    create_table(:"public.tenants") do |t|
+      t.string :name
+    end
   end
 end
 ActiveRecord::Migration.verbose = false
 CreateAllTables.up
+